requirements.txt
ipython==8.0.1
matplotlib==3.4.3
numpy==1.20.3
Pillow==9.0.0
pixellib==0.7.1
streamlit==1.5.0
